+++ R/data.R
@@ -1,0 +1,11 @@
+#' Keating waveform, 1999 format
+#'
+#' This is a waveform capture from Schneider Electric's Victoria-Keating facility.
+#'
+#' @format a comtrade object, a list with two members,
+#'   generated using [`comtrade()`]
+#' \describe{
+#'   \item{config}{comtrade config object}
+#'   \item{data}{cpmtrade data object}
+#' }
+"keating_1999"