+++ R/tidyverse.R
@@ -8,5 +8,6 @@
 #' @rdname pipe
 #' @export
 #' @param lhs,rhs An object and a function to apply to it
+#' @keywords internal
 #'
 NULL