+++ tests/testthat/test-data.R
@@ -3,15 +3,19 @@ context("data")
 library("dplyr")
 
 config <- ct_example("keating_1999.CFG") %>% ct_read_config()
-data_full <-
+config[["sampling_rate"]][["endsamp"]] <- 4
+
+data <-
   ct_example("keating_1999.DAT") %>%
   ct_read_data(
     config[["##A"]],
     config[["##D"]],
     max(config$sampling_rate$endsamp)
   )
+attr(data, "spec") <- NULL
 
-data_top <- data_full[(1:4), ]
+data_config <- ct_example("keating_1999.DAT") %>% ct_read_data_config(config)
+attr(data_config, "spec") <- NULL
 
 data_sample <- tribble(
   ~n, ~timestamp, ~a000001, ~a000002, ~a000003, ~a000004, ~a000005, ~a000006, ~a000007, ~a000008,
@@ -22,5 +26,6 @@ data_sample <- tribble(
 )
 
 test_that("data works", {
-  expect_identical(data_top, data_sample)
+  expect_identical(data, data_sample)
+  expect_identical(data_config, data_sample)
 })