+++ NEWS.md
@@ -1,6 +1,18 @@
 # ieeecomtrade 0.0.0.9000
 
-* Adds sample external data `system.file("extdata", "comtrade_1999_ascii", package = "ieeecomtrade")`
+* Adds `vignette("config")` to detail the structure of the configuation object.
+
+* Adds function read a zip-file into a comtrade object, `ct_read_zip()`.
+
+* Adds function to convert a comtrade object to a data-frame with time and scaled values, `ct_instant()`. 
+
+* Adds constructor function for comtrade object, `comtrade()`.
+
+* Adds functions to read data files, `ct_read_data()` and `ct_read_data_config()`. 
+
+* Adds function to read configuration files, `ct_read_config()`.
+
+* Adds sample external data which can be accessed using the function `ct_example()`.
 
 * Adds `NEWS.md` file to track changes to the package.